+++ R/mod_alcc.R
@@ -266,13 +266,19 @@ mod_alcc <- function(data=NULL,
     # ALCC curve
     ggplot2::geom_path(data = curve, ggplot2::aes(x=fitted_STV,y=new_RY),
                        color="grey15", linewidth = 1) +
-    ggplot2::scale_y_continuous(limits = c(0, max(RY)),breaks=seq(0,max(RY)*2,10)) +
+    ggplot2::scale_y_continuous(limits = c(0, max(RY)),breaks=seq(0,max(RY)*2, 10)) +
+    # Giving more flexibility to x scale
     ggplot2::scale_x_continuous(
       breaks = seq(0, maxx,
-                   by = ifelse(maxx > 300, 30,
-                               ifelse(maxx > 200, 20,
-                                      ifelse(maxx > 100, 10, 
-                                             ifelse(maxx > 10, 2, 0.5)))))) +
+                   by = ifelse(maxx >= 300, 50,
+                         ifelse(maxx >= 200, 20,
+                          ifelse(maxx >= 100, 10, 
+                           ifelse(maxx >= 50, 5,
+                            ifelse(maxx >= 20, 2,
+                             ifelse(maxx >= 10, 1,
+                              ifelse(maxx >= 5, 0.5,
+                               ifelse(maxx >= 1, 0.2, 
+                                      0.1))))))))) ) +
     # Text annotations
     ggplot2::annotate("text",label = paste("CSTV =", round(CSTV,1), "ppm"),
                       x = CSTV, y = 0, angle = 90, hjust = 0, vjust = 1.5, col = "grey25") +